Потенциально рабочий бот

+++ cogs/team_lobby.py
@@ -7,7 +7,6 @@ from constants import MAX_ITEMS_IN_MENU, MAX_SELECT_MENUS
 import enum
 import json
 from models.room_author_settings import AuthorSettings
-from models.sync_member_in_db import register_member_in_guild
 
 
 class ChannelActions(str, enum.Enum):
@@ -282,35 +281,31 @@ class Lobby(commands.Cog):
 
     # region Get
 
-    async def get_text_channel_id(self, voice_channel,
-                                  voice_creator=False):
-        if voice_creator:
-            query = ("SELECT id "
-                     "FROM lobby_text_channel "
-                     "WHERE voice_channel_creator_id = $1 and guild_id = $2")
-            text_channel_id = await self.pool.fetchval(query, voice_channel.id,
-                                                       voice_channel.guild.id)
-        else:
-            query = ("SELECT text_channel_id "
-                     "FROM lobby_message "
-                     "WHERE voice_channel_id = $1 and guild_id = $2")
-            text_channel_id = await self.pool.fetchval(query, voice_channel.id,
-                                                       voice_channel.guild.id)
-        return text_channel_id
+    async def get_channel_creator_id(self, lobby_voice_channel_id):
+        query = ("SELECT voice_creator_id "
+                 "FROM lobby_created_voice_channel "
+                 "WHERE id = $1")
+        return await self.pool.fetchval(query, lobby_voice_channel_id)
+
+    async def get_text_channel_id(self, voice_channel_creator_id):
+        query = ("SELECT text_channel_id "
+                 "FROM lobby_voice_channel_creator_settings "
+                 "WHERE id = $1")
+        return await self.pool.fetchval(query, voice_channel_creator_id)
 
     async def get_lobby_message_id(self, voice_channel):
         query = ("SELECT id "
                  "FROM lobby_message "
-                 "WHERE voice_channel_id = $1 and guild_id = $2")
-        message_id = await self.pool.fetchval(query, voice_channel.id,
-                                              voice_channel.guild.id)
+                 "WHERE voice_channel_id = $1")
+        message_id = await self.pool.fetchval(query, voice_channel.id)
         return message_id
 
     async def get_lobby_info_message(self, voice_channel):
         message_id = await self.get_lobby_message_id(voice_channel)
         if not message_id:
             return
-        text_channel_id = await self.get_text_channel_id(voice_channel)
+        channel_creator_id = await self.get_channel_creator_id(voice_channel.id)
+        text_channel_id = await self.get_text_channel_id(channel_creator_id)
         if not text_channel_id:
             return
         message = await self.get_message_from_discord(voice_channel, message_id,
@@ -360,9 +355,8 @@ class Lobby(commands.Cog):
     async def get_voice_channel_author_id(self, voice_channel):
         get_channel_author_id = "SELECT user_id " \
                                 "FROM lobby_voice_channel_author " \
-                                "WHERE voice_channel_id = $1 and guild_id = $2"
-        channel_author_id = await self.pool.fetchval(get_channel_author_id, voice_channel.id,
-                                                     voice_channel.guild.id)
+                                "WHERE voice_channel_id = $1"
+        channel_author_id = await self.pool.fetchval(get_channel_author_id, voice_channel.id)
         return channel_author_id
 
     async def get_channel_overwrites(self, category,
@@ -436,17 +430,17 @@ class Lobby(commands.Cog):
                  "WHERE id = $1")
         return await self.pool.fetchval(query, int(voice_channel_id))
 
-    async def save_sent_lobby_info_to_db(self, voice_channel,
-                                         message, text_channel_id):
-        query = ("INSERT INTO lobby_message (id, voice_channel_id, text_channel_id)"
-                 "VALUES ($1, $2, $3)")
+    async def save_message_id_to_db(self, voice_channel,
+                                    message):
+        query = ("INSERT INTO lobby_message (id, voice_channel_id)"
+                 "VALUES ($1, $2)")
         await self.pool.execute(query, message.id,
-                                voice_channel.id, text_channel_id)
+                                voice_channel.id)
 
-    async def delete_sent_lobby_info_from_db(self, voice_channel):
+    async def delete_message_id_from_db(self, message_id):
         query = ("DELETE FROM lobby_message "
-                 "WHERE voice_channel_id = $1")
-        await self.pool.execute(query, voice_channel.id)
+                 "WHERE id = $1")
+        await self.pool.execute(query, message_id)
 
     @staticmethod
     async def update_lobby_info_message(message, voice_channel):
@@ -473,15 +467,22 @@ class Lobby(commands.Cog):
 
     async def set_voice_channel_author_id(self, member,
                                           voice_channel):
-        query = ("INSERT INTO lobby_voice_channel_author(voice_channel_id, guild_id, user_id) "
-                 "VALUES ($1, $2, $3)")
+        query = ("INSERT INTO lobby_voice_channel_author(voice_channel_id, user_id) "
+                 "VALUES ($1, $2)")
         await self.pool.execute(query, voice_channel.id,
-                                voice_channel.guild.id, member.id)
+                                member.id)
 
     async def delete_voice_channel_author_id(self, voice_channel):
         query = ("DELETE FROM lobby_voice_channel_author "
                  "WHERE voice_channel_id = $1")
         await self.pool.execute(query, voice_channel.id)
+
+    async def add_lobby_channel_to_db(self, created_channel_id,
+                                      channel_creator_id):
+        query = ("INSERT INTO lobby_created_voice_channel (id, voice_creator_id) "
+                 "VALUES ($1, $2)")
+        await self.pool.execute(query, created_channel_id,
+                                channel_creator_id)
 
     async def create_voice_channel(self, member,
                                    category, overwrites,
@@ -527,10 +528,11 @@ class Lobby(commands.Cog):
             return voice_channel
 
     async def is_lobby_category(self, channel):
+        channel_creator_id = await self.get_channel_creator_id(channel.id)
         query = ("SELECT category_id_for_new_channel "
                  "FROM lobby_voice_channel_creator_settings "
-                 "WHERE guild_id = $1 and category_id_for_new_channel = $2")
-        lobby_category_id = await self.pool.fetchval(query, channel.guild.id,
+                 "WHERE id = $1 and category_id_for_new_channel = $2")
+        lobby_category_id = await self.pool.fetchval(query, channel_creator_id,
                                                      channel.category.id)
         return True if lobby_category_id else False
 
@@ -550,16 +552,21 @@ class Lobby(commands.Cog):
                 print("Left lobby room")
                 if not before.channel.members:
                     await before.channel.edit(overwrites=temp_overwrites)
+                    query = ("SELECT voice_creator_id "
+                             "FROM lobby_created_voice_channel "
+                             "WHERE id = $1")
+                    voice_creator_id = await self.pool.fetchval(query, before.channel.id)
+                    if not voice_creator_id:
+                        return
                     query = ("SELECT custom "
                              "FROM lobby_voice_channel_creator_settings "
-                             "WHERE guild_id = $1 and category_id_for_new_channel = $2")
-                    custom = await self.pool.fetchval(query, before.channel.guild.id,
-                                                      before.channel.category.id)
+                             "WHERE id = $1")
+                    custom = await self.pool.fetchval(query, voice_creator_id)
                     print("Before channel is empty. Deleting")
                     if custom is False:
                         message = await self.get_lobby_info_message(before.channel)
                         if message:
-                            await self.delete_sent_lobby_info_from_db(message)
+                            await self.delete_message_id_from_db(message.id)
                             await message.delete()
                             print("Lobby info message deleted")
 
@@ -584,8 +591,6 @@ class Lobby(commands.Cog):
             if is_voice_creator:
                 voice_creator_id, custom = is_voice_creator
                 if voice_creator_id:
-                    await register_member_in_guild(self.pool, member.id,
-                                                   guild_id)
                     print("Voice creator")
                     query = ("SELECT category_id_for_new_channel "
                              "FROM lobby_voice_channel_creator_settings "
@@ -620,6 +625,7 @@ class Lobby(commands.Cog):
                             return
                         await asyncio.sleep(1)
                         if voice_channel.members:
+                            await self.add_lobby_channel_to_db(voice_channel.id, voice_creator_id)
                             await self.set_voice_channel_author_id(member, voice_channel)
                             buttons = CustomChannelDashboardButtons(self.pool, self.bot)
                             await voice_channel.send(embed=hello_embed, view=buttons)
@@ -645,6 +651,7 @@ class Lobby(commands.Cog):
                             return
                         await asyncio.sleep(1)
                         if voice_channel.members:
+                            await self.add_lobby_channel_to_db(voice_channel.id, voice_creator_id)
                             await self.set_voice_channel_author_id(member, voice_channel)
                             if required_role == RequestedRole.missing:
                                 query = ("SELECT role_not_found_message "
@@ -661,18 +668,17 @@ class Lobby(commands.Cog):
                             embed = self.create_lobby_info_embed(member, required_role,
                                                                  voice_channel, voice_channel.user_limit)
 
-                            text_channel_id = await self.get_text_channel_id(channel_creator, voice_creator=True)
+                            text_channel_id = await self.get_text_channel_id(channel_creator.id)
                             text_channel = member.guild.get_channel(text_channel_id)
                             lobby_log_needed = await self.log_needed(channel_creator.id)
                             await voice_channel.edit(overwrites=overwrites)
                             if lobby_log_needed and not text_channel:
                                 await voice_channel.send(f"{member.mention}, "
-                                                         f"Оповещение о создании комнаты не было создано из-за неверных "
-                                                         f"настроек. Обратитесь к администратору за помощью")
+                                                         f"Оповещение о создании комнаты не было создано из-за неверных"
+                                                         f" настроек. Обратитесь к администратору за помощью")
                             elif text_channel:
                                 lobby_info_message = await text_channel.send(embed=embed)
-                                await self.save_sent_lobby_info_to_db(voice_channel, lobby_info_message,
-                                                                      text_channel.id)
+                                await self.save_message_id_to_db(voice_channel, lobby_info_message)
                             buttons = BaseDashboardButtons(self.pool, self.bot)
                             await voice_channel.send(embed=hello_embed, view=buttons)
                     if voice_channel and not voice_channel.members:
